+++ source/git/gitsource.go
@@ -36,6 +36,7 @@ import (
 )
 
 var validHex = regexp.MustCompile(`^[a-f0-9]{40}$`)
+var defaultBranch = regexp.MustCompile(`refs/heads/(\S+)`)
 
 type Opt struct {
 	CacheAccessor cache.Accessor
@@ -303,18 +304,8 @@ func (gs *gitSourceHandler) mountKnownHosts(ctx context.Context) (string, func()
 
 func (gs *gitSourceHandler) CacheKey(ctx context.Context, g session.Group, index int) (string, solver.CacheOpts, bool, error) {
 	remote := gs.src.Remote
-	ref := gs.src.Ref
-	if ref == "" {
-		ref = "master"
-	}
 	gs.locker.Lock(remote)
 	defer gs.locker.Unlock(remote)
-
-	if isCommitSHA(ref) {
-		ref = gs.shaToCacheKey(ref)
-		gs.cacheKey = ref
-		return ref, nil, true, nil
-	}
 
 	gs.getAuthToken(ctx, g)
 
@@ -344,6 +335,20 @@ func (gs *gitSourceHandler) CacheKey(ctx context.Context, g session.Group, index
 		defer unmountKnownHosts()
 	}
 
+	ref := gs.src.Ref
+	if ref == "" {
+		ref, err = getDefaultBranch(ctx, gitDir, "", sock, knownHosts, gs.auth, gs.src.Remote)
+		if err != nil {
+			return "", nil, false, err
+		}
+	}
+
+	if isCommitSHA(ref) {
+		ref = gs.shaToCacheKey(ref)
+		gs.cacheKey = ref
+		return ref, nil, true, nil
+	}
+
 	// TODO: should we assume that remote tag is immutable? add a timer?
 
 	buf, err := gitWithinDir(ctx, gitDir, "", sock, knownHosts, gs.auth, "ls-remote", "origin", ref)
@@ -366,11 +371,6 @@ func (gs *gitSourceHandler) CacheKey(ctx context.Context, g session.Group, index
 }
 
 func (gs *gitSourceHandler) Snapshot(ctx context.Context, g session.Group) (out cache.ImmutableRef, retErr error) {
-	ref := gs.src.Ref
-	if ref == "" {
-		ref = "master"
-	}
-
 	cacheKey := gs.cacheKey
 	if cacheKey == "" {
 		var err error
@@ -420,6 +420,14 @@ func (gs *gitSourceHandler) Snapshot(ctx context.Context, g session.Group) (out 
 			return nil, err
 		}
 		defer unmountKnownHosts()
+	}
+
+	ref := gs.src.Ref
+	if ref == "" {
+		ref, err = getDefaultBranch(ctx, gitDir, "", sock, knownHosts, gs.auth, gs.src.Remote)
+		if err != nil {
+			return nil, err
+		}
 	}
 
 	doFetch := true
@@ -641,4 +649,18 @@ func tokenScope(remote string) string {
 		}
 	}
 	return remote
+}
+
+// getDefaultBranch gets the default branch of a repository using ls-remote
+func getDefaultBranch(ctx context.Context, gitDir, workDir, sshAuthSock, knownHosts string, auth []string, remoteURL string) (string, error) {
+	buf, err := gitWithinDir(ctx, gitDir, workDir, sshAuthSock, knownHosts, auth, "ls-remote", "--symref", remoteURL, "HEAD")
+	if err != nil {
+		return "", errors.Errorf("error fetching default branch for repository %s: %v", remoteURL, err)
+	}
+
+	ss := defaultBranch.FindAllStringSubmatch(buf.String(), -1)
+	if len(ss) == 0 || len(ss[0]) != 2 {
+		return "", errors.Errorf("could not find default branch for repository: %s", remoteURL)
+	}
+	return ss[0][1], nil
 }
