+++ source/gitidentifier.go
@@ -2,8 +2,6 @@ package source
 
 import (
 	"net/url"
-	"os/exec"
-	"regexp"
 	"strings"
 
 	"github.com/moby/buildkit/util/sshutil"
@@ -20,8 +18,6 @@ type GitIdentifier struct {
 	MountSSHSock     string
 	KnownSSHHosts    string
 }
-
-var defaultBranch = regexp.MustCompile(`refs/heads/(\S+)`)
 
 func NewGitIdentifier(remoteURL string) (*GitIdentifier, error) {
 	repo := GitIdentifier{}
@@ -55,14 +51,6 @@ func NewGitIdentifier(remoteURL string) (*GitIdentifier, error) {
 		return nil, errors.Errorf("subdir not supported yet")
 	}
 
-	if repo.Ref == "" {
-		var err error
-		repo.Ref, err = getDefaultBranch(repo.Remote)
-		if err != nil {
-			return nil, err
-		}
-	}
-
 	return &repo, nil
 }
 
@@ -86,18 +74,4 @@ func getRefAndSubdir(fragment string) (ref string, subdir string) {
 		subdir = refAndDir[1]
 	}
 	return
-}
-
-// getDefaultBranch gets the default branch of a repository using ls-remote
-func getDefaultBranch(remoteURL string) (string, error) {
-	out, err := exec.Command("git", "ls-remote", "--symref", remoteURL, "HEAD").CombinedOutput()
-	if err != nil {
-		return "", errors.Errorf("error fetching default branch for repository %s: %v", remoteURL, err)
-	}
-
-	ss := defaultBranch.FindAllStringSubmatch(string(out), -1)
-	if len(ss) == 0 || len(ss[0]) != 2 {
-		return "", errors.Errorf("could not find default branch for repository: %s", remoteURL)
-	}
-	return ss[0][1], nil
 }
